Adding files to the index again
I add the files again because they have not been added
to the remote repository. I removed them because each time the
application was launched, changes were made to them in the editor,
which sometimes caused complications. However, I thought they would
be sent to the remote repository despite not being tracked.
Now I know that I do not.